refactor: qc-core point 도메인 패키지 정리
