+++ qc-application/src/main/java/com/eager/questioncloud/application/api/payment/ChargePointPaymentApprover.java
@@ -1,8 +1,8 @@
 package com.eager.questioncloud.application.api.payment;
 
-import com.eager.questioncloud.core.domain.point.ChargePointPayment;
-import com.eager.questioncloud.core.domain.point.ChargePointPaymentRepository;
-import com.eager.questioncloud.core.domain.point.PGPayment;
+import com.eager.questioncloud.core.domain.point.dto.PGPayment;
+import com.eager.questioncloud.core.domain.point.infrastructure.ChargePointPaymentRepository;
+import com.eager.questioncloud.core.domain.point.model.ChargePointPayment;
 import com.eager.questioncloud.exception.CustomException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
