+++ qc-application/src/main/java/com/eager/questioncloud/application/api/payment/ChargePointPaymentFailHandler.java
@@ -1,8 +1,8 @@
 package com.eager.questioncloud.application.api.payment;
 
 
-import com.eager.questioncloud.core.domain.point.ChargePointPayment;
-import com.eager.questioncloud.core.domain.point.ChargePointPaymentRepository;
+import com.eager.questioncloud.core.domain.point.infrastructure.ChargePointPaymentRepository;
+import com.eager.questioncloud.core.domain.point.model.ChargePointPayment;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
