+++ qc-application/src/main/java/com/eager/questioncloud/application/api/payment/PGAPI.java
@@ -1,7 +1,7 @@
 package com.eager.questioncloud.application.api.payment;
 
 
-import com.eager.questioncloud.core.domain.point.PGPayment;
+import com.eager.questioncloud.core.domain.point.dto.PGPayment;
 import com.eager.questioncloud.pg.portone.PortoneAPI;
 import com.eager.questioncloud.pg.portone.PortonePayment;
 import lombok.RequiredArgsConstructor;
