+++ qc-application/src/main/java/com/eager/questioncloud/application/api/payment/QuestionPaymentProcessor.java
@@ -4,11 +4,11 @@ import com.eager.questioncloud.core.domain.coupon.infrastructure.CouponRepositor
 import com.eager.questioncloud.core.domain.coupon.infrastructure.UserCouponRepository;
 import com.eager.questioncloud.core.domain.coupon.model.Coupon;
 import com.eager.questioncloud.core.domain.coupon.model.UserCoupon;
-import com.eager.questioncloud.core.domain.payment.model.QuestionPayment;
 import com.eager.questioncloud.core.domain.payment.infrastructure.QuestionPaymentOrderRepository;
 import com.eager.questioncloud.core.domain.payment.infrastructure.QuestionPaymentRepository;
-import com.eager.questioncloud.core.domain.point.UserPoint;
-import com.eager.questioncloud.core.domain.point.UserPointRepository;
+import com.eager.questioncloud.core.domain.payment.model.QuestionPayment;
+import com.eager.questioncloud.core.domain.point.infrastructure.UserPointRepository;
+import com.eager.questioncloud.core.domain.point.model.UserPoint;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
