+++ qc-application/src/main/java/com/eager/questioncloud/application/api/user/point/UserPointService.java
@@ -1,7 +1,7 @@
 package com.eager.questioncloud.application.api.user.point;
 
-import com.eager.questioncloud.core.domain.point.UserPoint;
-import com.eager.questioncloud.core.domain.point.UserPointRepository;
+import com.eager.questioncloud.core.domain.point.infrastructure.UserPointRepository;
+import com.eager.questioncloud.core.domain.point.model.UserPoint;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
