+++ qc-application/src/main/java/com/eager/questioncloud/application/exception/FailChargePointExceptionHandler.java
@@ -1,8 +1,8 @@
 package com.eager.questioncloud.application.exception;
 
 import com.eager.questioncloud.application.api.payment.PGAPI;
-import com.eager.questioncloud.core.domain.point.ChargePointPayment;
-import com.eager.questioncloud.core.domain.point.ChargePointPaymentRepository;
+import com.eager.questioncloud.core.domain.point.infrastructure.ChargePointPaymentRepository;
+import com.eager.questioncloud.core.domain.point.model.ChargePointPayment;
 import com.eager.questioncloud.exception.CustomException;
 import com.eager.questioncloud.exception.Error;
 import com.eager.questioncloud.exception.FailChargePointException;
