+++ qc-application/src/main/java/com/eager/questioncloud/application/exception/FailQuestionPaymentExceptionHandler.java
@@ -2,10 +2,10 @@ package com.eager.questioncloud.application.exception;
 
 import com.eager.questioncloud.core.domain.coupon.infrastructure.UserCouponRepository;
 import com.eager.questioncloud.core.domain.coupon.model.UserCoupon;
-import com.eager.questioncloud.core.domain.payment.model.QuestionPayment;
 import com.eager.questioncloud.core.domain.payment.infrastructure.QuestionPaymentRepository;
-import com.eager.questioncloud.core.domain.point.UserPoint;
-import com.eager.questioncloud.core.domain.point.UserPointRepository;
+import com.eager.questioncloud.core.domain.payment.model.QuestionPayment;
+import com.eager.questioncloud.core.domain.point.infrastructure.UserPointRepository;
+import com.eager.questioncloud.core.domain.point.model.UserPoint;
 import com.eager.questioncloud.exception.CustomException;
 import com.eager.questioncloud.exception.Error;
 import com.eager.questioncloud.exception.FailQuestionPaymentException;
