+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/point/dto/PGPayment.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.core.domain.point;
+package com.eager.questioncloud.core.domain.point.dto;
 
 import lombok.Getter;
