+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/point/implement/UserPointUpdater.java
@@ -1,5 +1,8 @@
-package com.eager.questioncloud.core.domain.point;
+package com.eager.questioncloud.core.domain.point.implement;
 
+import com.eager.questioncloud.core.domain.point.event.ChargePointEvent;
+import com.eager.questioncloud.core.domain.point.infrastructure.UserPointRepository;
+import com.eager.questioncloud.core.domain.point.model.UserPoint;
 import com.eager.questioncloud.exception.FailChargePointException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.event.EventListener;
