+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/point/infrastructure/ChargePointPaymentRepository.java
@@ -1,6 +1,7 @@
-package com.eager.questioncloud.core.domain.point;
+package com.eager.questioncloud.core.domain.point.infrastructure;
 
 import com.eager.questioncloud.core.common.PagingInformation;
+import com.eager.questioncloud.core.domain.point.model.ChargePointPayment;
 import java.util.List;
 
 public interface ChargePointPaymentRepository {
