+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/point/infrastructure/UserPointEntity.java
@@ -1,5 +1,6 @@
-package com.eager.questioncloud.core.domain.point;
+package com.eager.questioncloud.core.domain.point.infrastructure;
 
+import com.eager.questioncloud.core.domain.point.model.UserPoint;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
